backend/requirements.txt
Flask-Cors==3.0.4
Flask-RESTful==0.3.6
keras==2.0.2
requests==2.18.4
pillow==4.3.0
tensorflow
h5py==2.7.1